chore: move JvmOnlyRecord.java to the extension's runtime module
